Moved Over Pickup Methods From Board To PickupManager
