+++ Assets/Scripts/Managers/PickupManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 46a0ea7c120401f4cbb70cfc190e119f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
