+++ UserSettings/EditorUserSettings.asset
@@ -20,14 +20,8 @@ EditorUserSettings:
     RecentlyUsedSceneGuid-4:
       value: 50555551530d085d55565c7b47750b44464f1e7e7a7873652c794a36e4e16c6e
       flags: 0
-    RecentlyUsedSceneGuid-2:
-      value: 07520d5403050b5e0b080d7511770843474e1c7d7c2d71677e2a4a65bab06539
-      flags: 0
-    RecentlyUsedSceneGuid-3:
-      value: 5653015152535e0e0e5c5b2116775b464115492b2a787f69287d446ae7e23061
-      flags: 0
-    RecentlyUsedSceneGuid-4:
-      value: 5409065553075003085e5e7447205b4446151e7c2e2b73367b714a32b3b13261
+    RecentlyUsedSceneGuid-5:
+      value: 5353525f5201080a5c0a5a2415725e44121549737e7175697b2f1f66b1b63639
       flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
